requirements.txt
notion-client
jinja2
weasyprint
python-dotenv
